+++ xLights/BufferPanel.cpp
@@ -1038,7 +1038,7 @@ void BufferPanel::OnBufferStyleChoiceSelect(wxCommandEvent& event)
 
         // load the camera positions
         xLightsFrame* frame = xLightsApp::GetFrame();
-        for (size_t i = 0; i < frame->viewpoint_mgr.GetNum3DCameras(); ++i)
+        for (int i = 0; i < frame->viewpoint_mgr.GetNum3DCameras(); ++i)
         {
             Choice_PerPreviewCamera->Append(frame->viewpoint_mgr.GetCamera3D(i)->GetName());
         }
